tools/telegram/lee-grieg-telegram-connector/lee-telegram-connector/services/worker/requirements.txt
redis==5.0.8
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
python-dotenv==1.0.1
